project_b_rag_chatbot/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
google-genai>=1.0.0
scikit-learn>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
